+++ tests/test_dataframe/test_get.py
@@ -5,7 +5,7 @@ import pytest
 
 def test_get_cell():
     actual = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9]}, index=[10, 11, 13], columns=['a', 'b', 'c'],
-                          sorted=False)
+                          sort=False)
 
     assert actual.get(10, 'a') == 1
     assert actual.get(11, 'a') == 2
@@ -24,7 +24,7 @@ def test_get_cell():
 
 def test_get_cell_sorted():
     actual = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9]}, index=[10, 11, 13], columns=['a', 'b', 'c'],
-                          sorted=True)
+                          sort=True)
 
     assert actual.get(10, 'a') == 1
     assert actual.get(11, 'a') == 2
@@ -43,9 +43,9 @@ def test_get_cell_sorted():
 
 def test_get_rows():
     df = rc.DataFrame({'a': [1, 2, 3, 4], 'b': [4, 5, 6, 7], 'c': [7, 8, 9, None]}, index=[10, 11, 12, 99],
-                      columns=['a', 'b', 'c'], index_name='start_10', sorted=False)
+                      columns=['a', 'b', 'c'], index_name='start_10', sort=False)
 
-    expected = rc.DataFrame({'c': [8, 9]}, index=[11, 12], index_name='start_10', sorted=False)
+    expected = rc.DataFrame({'c': [8, 9]}, index=[11, 12], index_name='start_10', sort=False)
     actual = df.get([11, 12], 'c')
     assert_frame_equal(actual, expected)
 
@@ -54,7 +54,7 @@ def test_get_rows():
     assert_frame_equal(actual, expected)
 
     # index out of order
-    expected = rc.DataFrame({'c': [None, 7]}, index=[99, 10], index_name='start_10', sorted=False)
+    expected = rc.DataFrame({'c': [None, 7]}, index=[99, 10], index_name='start_10', sort=False)
     actual = df.get([99, 10], 'c')
     assert_frame_equal(actual, expected)
 
@@ -78,9 +78,9 @@ def test_get_rows():
 
 def test_get_rows_sorted():
     df = rc.DataFrame({'a': [1, 2, 3, 4], 'b': [4, 5, 6, 7], 'c': [7, 8, 9, None]}, index=[10, 11, 12, 99],
-                      columns=['a', 'b', 'c'], index_name='start_10', sorted=True)
+                      columns=['a', 'b', 'c'], index_name='start_10', sort=True)
 
-    expected = rc.DataFrame({'c': [8, 9]}, index=[11, 12], index_name='start_10', sorted=True)
+    expected = rc.DataFrame({'c': [8, 9]}, index=[11, 12], index_name='start_10', sort=True)
     actual = df.get([11, 12], 'c')
     assert_frame_equal(actual, expected)
 
@@ -92,7 +92,7 @@ def test_get_rows_sorted():
     assert_frame_equal(actual, expected)
 
     # index out of order
-    expected = rc.DataFrame({'c': [7, None]}, index=[10, 99], index_name='start_10', sorted=True)
+    expected = rc.DataFrame({'c': [7, None]}, index=[10, 99], index_name='start_10', sort=True)
     actual = df.get([99, 10], 'c')
     assert_frame_equal(actual, expected)
 
@@ -113,10 +113,10 @@ def test_get_rows_sorted():
 
 def test_get_columns():
     df = rc.DataFrame({'a': [1, 2, 3, 4], 'b': [4, 5, 6, 7], 'c': [7, 8, 9, None]}, index=[10, 11, 12, 99],
-                      columns=['a', 'b', 'c'], index_name='start_10', sorted=False)
+                      columns=['a', 'b', 'c'], index_name='start_10', sort=False)
 
     expected = rc.DataFrame({'a': [4], 'c': [None]}, index=[99], columns=['a', 'c'], index_name='start_10',
-                            sorted=False)
+                            sort=False)
     actual = df.get(99, ['a', 'c'])
     assert_frame_equal(actual, expected)
 
@@ -126,7 +126,7 @@ def test_get_columns():
 
     # columns out of order
     expected = rc.DataFrame({'c': [8], 'b': [5]}, index=[11], columns=['c', 'b'], index_name='start_10',
-                            sorted=False)
+                            sort=False)
     actual = df.get(11, ['c', 'b'])
     assert_frame_equal(actual, expected)
 
@@ -145,10 +145,10 @@ def test_get_columns():
 
 def test_get_columns_sorted():
     df = rc.DataFrame({'a': [1, 2, 3, 4], 'b': [4, 5, 6, 7], 'c': [7, 8, 9, None]}, index=[10, 11, 12, 99],
-                      columns=['a', 'b', 'c'], index_name='start_10', sorted=True)
+                      columns=['a', 'b', 'c'], index_name='start_10', sort=True)
 
     expected = rc.DataFrame({'a': [4], 'c': [None]}, index=[99], columns=['a', 'c'], index_name='start_10',
-                            sorted=True)
+                            sort=True)
     actual = df.get(99, ['a', 'c'])
     assert_frame_equal(actual, expected)
 
@@ -158,7 +158,7 @@ def test_get_columns_sorted():
 
     # columns out of order
     expected = rc.DataFrame({'c': [8], 'b': [5]}, index=[11], columns=['c', 'b'], index_name='start_10',
-                            sorted=True)
+                            sort=True)
     actual = df.get(11, ['c', 'b'])
     assert_frame_equal(actual, expected)
 
@@ -173,10 +173,10 @@ def test_get_columns_sorted():
 
 def test_get_matrix():
     df = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9], 'd': [10, 11, 12]}, index=['x', 'y', 'z'],
-                      columns=['a', 'b', 'c', 'd'], index_name='letters', sorted=False)
+                      columns=['a', 'b', 'c', 'd'], index_name='letters', sort=False)
 
     expected = rc.DataFrame({'b': [4, 6], 'd': [10, 12]}, index=['x', 'z'], columns=['b', 'd'], index_name='letters',
-                            sorted=False)
+                            sort=False)
     actual = df.get(['x', 'z'], ['b', 'd'])
     assert_frame_equal(actual, expected)
 
@@ -186,7 +186,7 @@ def test_get_matrix():
 
     # columns out of order
     expected = rc.DataFrame({'d': [10, 12], 'c': [7, 9]}, index=['x', 'z'], columns=['d', 'c'], index_name='letters',
-                            sorted=False)
+                            sort=False)
     actual = df.get(['x', 'z'], ['d', 'c'])
     assert_frame_equal(actual, expected)
 
@@ -213,10 +213,10 @@ def test_get_matrix():
 
 def test_get_matrix_sorted():
     df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6], 'c': [8, 7, 9], 'd': [11, 10, 12]}, index=['y', 'x', 'z'],
-                      columns=['a', 'b', 'c', 'd'], index_name='letters', sorted=True)
+                      columns=['a', 'b', 'c', 'd'], index_name='letters', sort=True)
 
     expected = rc.DataFrame({'b': [4, 6], 'd': [10, 12]}, index=['x', 'z'], columns=['b', 'd'], index_name='letters',
-                            sorted=True)
+                            sort=True)
     actual = df.get(['x', 'z'], ['b', 'd'])
     assert_frame_equal(actual, expected)
 
@@ -226,7 +226,7 @@ def test_get_matrix_sorted():
 
     # columns out of order
     expected = rc.DataFrame({'d': [10, 12], 'c': [7, 9]}, index=['x', 'z'], columns=['d', 'c'], index_name='letters',
-                            sorted=True)
+                            sort=True)
     actual = df.get(['x', 'z'], ['d', 'c'])
     assert_frame_equal(actual, expected)
 
@@ -291,16 +291,16 @@ def test_get_locations():
 
 def test_get_square_brackets():
     df = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9], 'd': [10, 11, 12]}, columns=['a', 'b', 'c', 'd'],
-                      sorted=False)
+                      sort=False)
 
     # df['b'] -- get column
-    assert_frame_equal(df['b'], rc.DataFrame({'b': [4, 5, 6]}, sorted=False))
+    assert_frame_equal(df['b'], rc.DataFrame({'b': [4, 5, 6]}, sort=False))
 
     # df[['a', 'b', c']] -- get columns
     assert_frame_equal(df[['a', 'b', 'c']], rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9]},
-                                                         columns=['a', 'b', 'c'], sorted=False))
+                                                         columns=['a', 'b', 'c'], sort=False))
 
-    assert_frame_equal(df[['c', 'a']], rc.DataFrame({'c': [7, 8, 9], 'a': [1, 2, 3]}, columns=['c', 'a'], sorted=False))
+    assert_frame_equal(df[['c', 'a']], rc.DataFrame({'c': [7, 8, 9], 'a': [1, 2, 3]}, columns=['c', 'a'], sort=False))
 
     # df[1, 'd'] -- get cell at index = 5, column = 'b'
     assert df[1, 'd'] == 11
@@ -308,37 +308,37 @@ def test_get_square_brackets():
     # df[[0, 2]] -- get indexes = [0, 2] all columns
     assert_frame_equal(df[[0, 2], df.columns],
                        rc.DataFrame({'a': [1, 3], 'b': [4, 6], 'c': [7, 9], 'd': [10, 12]},
-                                    columns=['a', 'b', 'c', 'd'], index=[0, 2], sorted=False))
+                                    columns=['a', 'b', 'c', 'd'], index=[0, 2], sort=False))
 
     assert_frame_equal(df[[2, 1], df.columns],
                        rc.DataFrame({'a': [3, 2], 'b': [6, 5], 'c': [9, 8], 'd': [12, 11]},
-                                    columns=['a', 'b', 'c', 'd'], index=[2, 1], sorted=False))
+                                    columns=['a', 'b', 'c', 'd'], index=[2, 1], sort=False))
 
     # df[[0, 2], 'c'] -- get indexes = [4, 5], column = 'b'
-    assert_frame_equal(df[[0, 2], 'c'], rc.DataFrame({'c': [7, 9]}, index=[0, 2], sorted=False))
+    assert_frame_equal(df[[0, 2], 'c'], rc.DataFrame({'c': [7, 9]}, index=[0, 2], sort=False))
 
-    assert_frame_equal(df[[2, 0], 'c'], rc.DataFrame({'c': [9, 7]}, index=[2, 0], sorted=False))
+    assert_frame_equal(df[[2, 0], 'c'], rc.DataFrame({'c': [9, 7]}, index=[2, 0], sort=False))
 
     # df[[1, 2], ['a', 'd']] -- get indexes = [4, 5], columns = ['a', 'b']
     assert_frame_equal(df[[1, 2], ['a', 'd']], rc.DataFrame({'a': [2, 3], 'd': [11, 12]}, columns=['a', 'd'],
-                       index=[1, 2], sorted=False))
+                                                            index=[1, 2], sort=False))
 
     assert_frame_equal(df[[2, 0], ['d', 'a']], rc.DataFrame({'d': [12, 10], 'a': [3, 1]}, columns=['d', 'a'],
-                       index=[2, 0], sorted=False))
+                                                            index=[2, 0], sort=False))
 
 
 def test_get_square_brackets_sorted():
     df = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9], 'd': [10, 11, 12]}, columns=['a', 'b', 'c', 'd'],
-                      sorted=True)
+                      sort=True)
 
     # df['b'] -- get column
-    assert_frame_equal(df['b'], rc.DataFrame({'b': [4, 5, 6]}, sorted=True))
+    assert_frame_equal(df['b'], rc.DataFrame({'b': [4, 5, 6]}, sort=True))
 
     # df[['a', 'b', c']] -- get columns
     assert_frame_equal(df[['a', 'b', 'c']], rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9]},
-                                                         columns=['a', 'b', 'c'], sorted=True))
+                                                         columns=['a', 'b', 'c'], sort=True))
 
-    assert_frame_equal(df[['c', 'a']], rc.DataFrame({'c': [7, 8, 9], 'a': [1, 2, 3]}, columns=['c', 'a'], sorted=True))
+    assert_frame_equal(df[['c', 'a']], rc.DataFrame({'c': [7, 8, 9], 'a': [1, 2, 3]}, columns=['c', 'a'], sort=True))
 
     # df[1, 'd'] -- get cell at index = 5, column = 'b'
     assert df[1, 'd'] == 11
@@ -346,43 +346,43 @@ def test_get_square_brackets_sorted():
     # df[[0, 2]] -- get indexes = [0, 2] all columns
     assert_frame_equal(df[[0, 2], df.columns],
                        rc.DataFrame({'a': [1, 3], 'b': [4, 6], 'c': [7, 9], 'd': [10, 12]},
-                                    columns=['a', 'b', 'c', 'd'], index=[0, 2], sorted=True))
+                                    columns=['a', 'b', 'c', 'd'], index=[0, 2], sort=True))
 
     assert_frame_equal(df[[2, 1], df.columns],
                        rc.DataFrame({'a': [2, 3], 'b': [5, 6], 'c': [8, 9], 'd': [11, 12]},
-                                    columns=['a', 'b', 'c', 'd'], index=[1, 2], sorted=True))
+                                    columns=['a', 'b', 'c', 'd'], index=[1, 2], sort=True))
 
     # df[[0, 2], 'c'] -- get indexes = [4, 5], column = 'b'
-    assert_frame_equal(df[[0, 2], 'c'], rc.DataFrame({'c': [7, 9]}, index=[0, 2], sorted=True))
+    assert_frame_equal(df[[0, 2], 'c'], rc.DataFrame({'c': [7, 9]}, index=[0, 2], sort=True))
 
-    assert_frame_equal(df[[2, 0], 'c'], rc.DataFrame({'c': [9, 7]}, index=[2, 0], sorted=True))
+    assert_frame_equal(df[[2, 0], 'c'], rc.DataFrame({'c': [9, 7]}, index=[2, 0], sort=True))
 
     # df[[1, 2], ['a', 'd']] -- get indexes = [4, 5], columns = ['a', 'b']
     assert_frame_equal(df[[1, 2], ['a', 'd']], rc.DataFrame({'a': [2, 3], 'd': [11, 12]}, columns=['a', 'd'],
-                       index=[1, 2], sorted=True))
+                                                            index=[1, 2], sort=True))
 
     assert_frame_equal(df[[2, 0], ['d', 'a']], rc.DataFrame({'d': [10, 12], 'a': [1, 3]}, columns=['d', 'a'],
-                       index=[0, 2], sorted=True))
+                                                            index=[0, 2], sort=True))
 
 
 def test_get_slicer():
     df = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9], 'd': [10, 11, 12]}, columns=['a', 'b', 'c', 'd'],
-                      sorted=False)
+                      sort=False)
 
     # df[1:2] -- get slice from index 1 to 2, all columns
     assert_frame_equal(df[1:2],
                        rc.DataFrame({'a': [2, 3], 'b': [5, 6], 'c': [8, 9], 'd': [11, 12]},
-                                    columns=['a', 'b', 'c', 'd'], index=[1, 2], sorted=False))
+                                    columns=['a', 'b', 'c', 'd'], index=[1, 2], sort=False))
 
     # df[0:1, ['c', 'd']] -- get slice from index 0 to 1, columns ['c', 'd']
     assert_frame_equal(df[0:1, ['c', 'd']], rc.DataFrame({'c': [7, 8], 'd': [10, 11]},
-                                                         columns=['c', 'd'], index=[0, 1], sorted=False))
+                                                         columns=['c', 'd'], index=[0, 1], sort=False))
 
     assert_frame_equal(df[0:1, ['d', 'c']], rc.DataFrame({'d': [10, 11], 'c': [7, 8]},
-                                                         columns=['d', 'c'], index=[0, 1], sorted=False))
+                                                         columns=['d', 'c'], index=[0, 1], sort=False))
 
     # df[1:1, 'c'] -- get slice 1 to 1 and column 'c'
-    assert_frame_equal(df[1:1, 'c'], rc.DataFrame({'c': [8]}, index=[1], sorted=False))
+    assert_frame_equal(df[1:1, 'c'], rc.DataFrame({'c': [8]}, index=[1], sort=False))
 
     # test indexes not in the range
     with pytest.raises(IndexError):
@@ -397,22 +397,22 @@ def test_get_slicer():
 
 def test_get_slicer_sorted():
     df = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': [7, 8, 9], 'd': [10, 11, 12]}, columns=['a', 'b', 'c', 'd'],
-                      sorted=True)
+                      sort=True)
 
     # df[1:2] -- get slice from index 1 to 2, all columns
     assert_frame_equal(df[1:2],
                        rc.DataFrame({'a': [2, 3], 'b': [5, 6], 'c': [8, 9], 'd': [11, 12]},
-                                    columns=['a', 'b', 'c', 'd'], index=[1, 2], sorted=True))
+                                    columns=['a', 'b', 'c', 'd'], index=[1, 2], sort=True))
 
     # df[0:1, ['c', 'd']] -- get slice from index 0 to 1, columns ['c', 'd']
     assert_frame_equal(df[0:1, ['c', 'd']], rc.DataFrame({'c': [7, 8], 'd': [10, 11]},
-                                                         columns=['c', 'd'], index=[0, 1], sorted=True))
+                                                         columns=['c', 'd'], index=[0, 1], sort=True))
 
     assert_frame_equal(df[0:1, ['d', 'c']], rc.DataFrame({'d': [10, 11], 'c': [7, 8]},
-                                                         columns=['d', 'c'], index=[0, 1], sorted=True))
+                                                         columns=['d', 'c'], index=[0, 1], sort=True))
 
     # df[1:1, 'c'] -- get slice 1 to 1 and column 'c'
-    assert_frame_equal(df[1:1, 'c'], rc.DataFrame({'c': [8]}, index=[1], sorted=True))
+    assert_frame_equal(df[1:1, 'c'], rc.DataFrame({'c': [8]}, index=[1], sort=True))
 
     # test indexes not in the range
     with pytest.raises(IndexError):
@@ -426,28 +426,28 @@ def test_get_slicer_sorted():
 
 
 def test_head():
-    df = rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sorted=False)
+    df = rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sort=False)
 
-    assert_frame_equal(df.head(0), rc.DataFrame(columns=[1, 2], sorted=False))
-    assert_frame_equal(df.head(1), rc.DataFrame({1: [0], 2: [3]}, columns=[1, 2], sorted=False))
-    assert_frame_equal(df.head(2), rc.DataFrame({1: [0, 1], 2: [3, 4]}, columns=[1, 2], sorted=False))
-    assert_frame_equal(df.head(3), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sorted=False))
-    assert_frame_equal(df.head(999), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sorted=False))
+    assert_frame_equal(df.head(0), rc.DataFrame(columns=[1, 2], sort=False))
+    assert_frame_equal(df.head(1), rc.DataFrame({1: [0], 2: [3]}, columns=[1, 2], sort=False))
+    assert_frame_equal(df.head(2), rc.DataFrame({1: [0, 1], 2: [3, 4]}, columns=[1, 2], sort=False))
+    assert_frame_equal(df.head(3), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sort=False))
+    assert_frame_equal(df.head(999), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sort=False))
 
 
 def test_tail():
-    df = rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sorted=False)
+    df = rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sort=False)
 
-    assert_frame_equal(df.tail(0), rc.DataFrame(columns=[1, 2], sorted=False))
-    assert_frame_equal(df.tail(1), rc.DataFrame({1: [2], 2: [5]}, columns=[1, 2], index=[2], sorted=False))
-    assert_frame_equal(df.tail(2), rc.DataFrame({1: [1, 2], 2: [4, 5]}, columns=[1, 2], index=[1, 2], sorted=False))
-    assert_frame_equal(df.tail(3), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sorted=False))
-    assert_frame_equal(df.tail(999), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sorted=False))
+    assert_frame_equal(df.tail(0), rc.DataFrame(columns=[1, 2], sort=False))
+    assert_frame_equal(df.tail(1), rc.DataFrame({1: [2], 2: [5]}, columns=[1, 2], index=[2], sort=False))
+    assert_frame_equal(df.tail(2), rc.DataFrame({1: [1, 2], 2: [4, 5]}, columns=[1, 2], index=[1, 2], sort=False))
+    assert_frame_equal(df.tail(3), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sort=False))
+    assert_frame_equal(df.tail(999), rc.DataFrame({1: [0, 1, 2], 2: [3, 4, 5]}, columns=[1, 2], sort=False))
 
 
 def test_get_index():
     df = rc.DataFrame({'a': [1, 2, 3, 4], 'b': [4, 5, 6, 7], 'c': [7, 8, 9, None]}, index=[10, 11, 12, 99],
-                      columns=['a', 'b', 'c'], index_name='start_10', sorted=False)
+                      columns=['a', 'b', 'c'], index_name='start_10', sort=False)
 
     assert df.get_index() == df.index