+++ tests/test_dataframe/test_initialize.py
@@ -12,14 +12,14 @@ def test_default_empty_init():
     assert actual.data == []
     assert actual.columns == []
     assert actual.index == []
-    assert actual.sorted is True
+    assert actual.sort is True
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
     assert all([isinstance(actual.data[x], list) for x in range(len(actual.columns))])
 
-    actual = rc.DataFrame(sorted=False)
-    assert actual.sorted is False
+    actual = rc.DataFrame(sort=False)
+    assert actual.sort is False
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -29,7 +29,7 @@ def test_default_empty_init():
     assert actual.data == [[], [], []]
     assert actual.columns == ['a', 'b', 'c']
     assert actual.index == []
-    assert actual.sorted is True
+    assert actual.sort is True
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -39,7 +39,7 @@ def test_default_empty_init():
     assert actual.data == [[None, None, None], [None, None, None]]
     assert actual.columns == ['a', 'b']
     assert actual.index == [1, 2, 3]
-    assert actual.sorted is False
+    assert actual.sort is False
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -49,14 +49,14 @@ def test_default_empty_init():
     assert actual.data == [[None, None, None], [None, None, None]]
     assert actual.columns == ['a', 'b']
     assert actual.index == [1, 2, 3]
-    assert actual.sorted is False
+    assert actual.sort is False
     assert isinstance(actual.index, blist)
     assert isinstance(actual.columns, blist)
     assert isinstance(actual.data, blist)
     assert all([isinstance(actual.data[x], blist) for x in range(len(actual.columns))])
 
-    actual = rc.DataFrame(index=[1, 2, 3], columns=['a', 'b'], sorted=True)
-    assert actual.sorted is True
+    actual = rc.DataFrame(index=[1, 2, 3], columns=['a', 'b'], sort=True)
+    assert actual.sort is True
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -69,7 +69,7 @@ def test_default_init():
     assert set(tuple(x) for x in actual.data) == {(1, 2, 3), (4, 5, 6)}
     assert set(actual.columns) == {'a', 'b'}
     assert actual.index == [0, 1, 2]
-    assert actual.sorted is True
+    assert actual.sort is True
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -81,7 +81,7 @@ def test_default_init():
     assert set(actual.columns) == {'a', 'b'}
     assert actual.index == ['a', 'b', 'c']
     assert actual.index_name == 'letters'
-    assert actual.sorted is False
+    assert actual.sort is False
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -92,7 +92,7 @@ def test_default_init():
     assert actual.data == [[4, 5, 6], [1, 2, 3]]
     assert actual.columns == ['b', 'a']
     assert actual.index == ['a', 'b', 'c']
-    assert actual.sorted is False
+    assert actual.sort is False
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -103,7 +103,7 @@ def test_default_init():
     assert actual.columns == ['b', 'c', 'a']
     assert actual.index == [0, 1, 2]
     assert actual.data == [[2, None, None], [1, 2, 3], [1, None, None]]
-    assert actual.sorted is True
+    assert actual.sort is True
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)
     assert isinstance(actual.data, list)
@@ -113,39 +113,39 @@ def test_default_init():
 def test_sorted_init():
     # initialized with index defaults to False
     df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], index=[12, 11, 13])
-    assert df.sorted is False
+    assert df.sort is False
 
-    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], index=[12, 11, 13], sorted=True)
-    assert df.sorted is True
+    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], index=[12, 11, 13], sort=True)
+    assert df.sort is True
     assert df.index == [11, 12, 13]
     assert df.data == [[1, 2, 3], [4, 5, 6]]
 
     # initialized with no index defaults to True
     df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'])
-    assert df.sorted is True
-    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], sorted=False)
-    assert df.sorted is False
+    assert df.sort is True
+    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], sort=False)
+    assert df.sort is False
 
-    # if sorted is true, but no index provided it will assume already in sorted order
-    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], sorted=True)
-    assert df.sorted is True
+    # if sort is true, but no index provided it will assume already in sort order
+    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], sort=True)
+    assert df.sort is True
     assert df.index == [0, 1, 2]
     assert df.data == [[2, 1, 3], [5, 4, 6]]
 
-    # start un-sorted, then set to sorted
-    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], index=[12, 11, 13], sorted=False)
-    assert df.sorted is False
+    # start un-sort, then set to sort
+    df = rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, columns=['a', 'b'], index=[12, 11, 13], sort=False)
+    assert df.sort is False
     assert df.index == [12, 11, 13]
     assert df.data == [[2, 1, 3], [5, 4, 6]]
 
-    df.sorted = True
+    df.sort = True
     assert df.index == [11, 12, 13]
     assert df.data == [[1, 2, 3], [4, 5, 6]]
 
-    # mixed type index will bork on sorted=True
+    # mixed type index will bork on sort=True
     if PYTHON3:
         with pytest.raises(TypeError):
-            rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, index=[1, 'b', 3], sorted=True)
+            rc.DataFrame({'a': [2, 1, 3], 'b': [5, 4, 6]}, index=[1, 'b', 3], sort=True)
 
 
 def test_jagged_data():
@@ -153,7 +153,7 @@ def test_jagged_data():
     assert actual.data == [[None, None, None], [1, None, None], [1, 2, None], [1, 2, 3]]
     assert actual.columns == ['a', 'b', 'c', 'd']
     assert actual.index == [0, 1, 2]
-    assert actual.sorted is True
+    assert actual.sort is True
     assert isinstance(actual.index, list)
     assert isinstance(actual.columns, list)