+++ tests/test_dataframe/test_utils.py
@@ -27,7 +27,7 @@ def test_assert_frame_equal():
     with pytest.raises(AssertionError):
         assert_frame_equal(df1, df2)
 
-    df2 = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]}, columns=['a', 'b'], index=[1, 2, 3], sorted=True)
+    df2 = rc.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]}, columns=['a', 'b'], index=[1, 2, 3], sort=True)
     with pytest.raises(AssertionError):
         assert_frame_equal(df1, df2)